4-chatbot/requirements-no-version.txt
aiohttp
aiosignal
annotated-types
anyio
appnope
asgiref
asttokens
async-timeout
attrs
backoff
bcrypt
build
cachetools
certifi
charset-normalizer
chroma-hnswlib
chromadb
click
coloredlogs
comm
dataclasses-json
debugpy
decorator
Deprecated
distro
docx2txt
durationpy
exceptiongroup
executing
fastapi
filelock
flatbuffers
frozenlist
fsspec
google-auth
googleapis-common-protos
grpcio
h11
httpcore
httptools
httpx
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
ipykernel
ipython
jedi
jiter
jsonpatch
jsonpointer
jupyter_client
jupyter_core
kubernetes
langchain
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-pinecone
langchain-text-splitters
langchain-upstage
langsmith
lxml
markdown-it-py
marshmallow
matplotlib-inline
mdurl
mmh3
monotonic
mpmath
multidict
mypy-extensions
nest-asyncio
numpy
oauthlib
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
parso
pexpect
pinecone
pinecone-plugin-interface
platformdirs
posthog
prompt_toolkit
propcache
protobuf
psutil
ptyprocess
pure_eval
pyasn1
pyasn1_modules
pydantic
pydantic-settings
pydantic_core
Pygments
pypdf
PyPika
pyproject_hooks
python-dateutil
python-docx
python-dotenv
PyYAML
pyzmq
regex
requests
requests-oauthlib
requests-toolbelt
rich
rsa
shellingham
six
sniffio
SQLAlchemy
stack-data
starlette
sympy
tenacity
tiktoken
tokenizers
tomli
tornado
tqdm
traitlets
typer
typing-inspect
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
wcwidth
websocket-client
websockets
wrapt
yarl
zipp